# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/operator_profiles:
    get:
      tags:
        - OperatorProfiles
      summary: List OperatorProfiles
      description: Get a paginated list of OperatorProfiles with optional filtering and sorting
      operationId: listOperatorProfiles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorProfileList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - OperatorProfiles
      summary: Create OperatorProfile
      description: Create a new OperatorProfile
      operationId: createOperatorProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperatorProfileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/sessions:
    get:
      tags:
        - Sessions
      summary: List Sessions
      description: Get a paginated list of Sessions with optional filtering and sorting
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Sessions
      summary: Create Session
      description: Create a new Session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    OperatorProfile:
      type: object
      required:
        - id
        - user_id
        - languages
        - metadata
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          nullable: true
          type: string
          maxLength: 200
        languages:
          type: array
        last_heartbeat_at:
          nullable: true
          type: string
          format: date-time
        last_assigned_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateOperatorProfileRequest:
      type: object
      required:
        - user_id
        - languages
        - metadata
      properties:
        user_id:
          type: string
          format: uuid
        display_name:
          nullable: true
          type: string
          maxLength: 200
        languages:
          type: array
        last_heartbeat_at:
          nullable: true
          type: string
          format: date-time
        last_assigned_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    OperatorProfileList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OperatorProfile'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Session:
      type: object
      required:
        - id
        - channel_id
        - failure
        - expertise_names
        - is_pending_request
        - message_count
        - last_interest_at
        - is_test
        - metadata
      properties:
        id:
          type: string
          format: uuid
        channel_id:
          type: string
          format: uuid
        title:
          nullable: true
          type: string
          maxLength: 300
        status:
          nullable: true
          type: string
        failure:
          type: string
        outcome:
          nullable: true
          type: string
        close_reason:
          nullable: true
          type: string
        closed_at:
          nullable: true
          type: string
          format: date-time
        operator_user_id:
          nullable: true
          type: string
          format: uuid
        chatbot_current_step_id:
          nullable: true
          type: string
          format: uuid
        expertise_names:
          type: array
        website_visitor_id:
          nullable: true
          type: string
          format: uuid
        visitor_country_code:
          nullable: true
          type: string
          maxLength: 2
        visitor_timezone:
          nullable: true
          type: string
          maxLength: 60
        is_pending_request:
          type: boolean
        crm_lead_id:
          nullable: true
          type: string
          format: uuid
        visitor_language:
          nullable: true
          type: string
          maxLength: 35
        message_count:
          type: integer
          format: int32
        first_response_at:
          nullable: true
          type: string
          format: date-time
        last_interest_at:
          type: string
          format: date-time
        last_visitor_message_at:
          nullable: true
          type: string
          format: date-time
        last_operator_message_at:
          nullable: true
          type: string
          format: date-time
        is_test:
          type: boolean
        error_detail:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    CreateSessionRequest:
      type: object
      required:
        - channel_id
        - failure
        - expertise_names
        - is_pending_request
        - message_count
        - last_interest_at
        - is_test
        - metadata
      properties:
        channel_id:
          type: string
          format: uuid
        title:
          nullable: true
          type: string
          maxLength: 300
        status:
          nullable: true
          type: string
        failure:
          type: string
        outcome:
          nullable: true
          type: string
        close_reason:
          nullable: true
          type: string
        closed_at:
          nullable: true
          type: string
          format: date-time
        operator_user_id:
          nullable: true
          type: string
          format: uuid
        chatbot_current_step_id:
          nullable: true
          type: string
          format: uuid
        expertise_names:
          type: array
        website_visitor_id:
          nullable: true
          type: string
          format: uuid
        visitor_country_code:
          nullable: true
          type: string
          maxLength: 2
        visitor_timezone:
          nullable: true
          type: string
          maxLength: 60
        is_pending_request:
          type: boolean
        crm_lead_id:
          nullable: true
          type: string
          format: uuid
        visitor_language:
          nullable: true
          type: string
          maxLength: 35
        message_count:
          type: integer
          format: int32
        first_response_at:
          nullable: true
          type: string
          format: date-time
        last_interest_at:
          type: string
          format: date-time
        last_visitor_message_at:
          nullable: true
          type: string
          format: date-time
        last_operator_message_at:
          nullable: true
          type: string
          format: date-time
        is_test:
          type: boolean
        error_detail:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    SessionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []